pyproject.toml
[project]
name = "ai-research-mcp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "fastapi-mcp>=0.3.7",
    "fastembed>=0.7.1",
    "google-cloud>=0.34.0",
    "google-cloud-storage>=3.2.0",
    "google-genai>=1.25.0",
    "pdfminer>=20191125",
    "pdfminer-six>=20250506",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.14.3",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
]
